feat : helper 클래스 사용하여 구조 변경
