+++ aiku/aiku-payment/src/main/java/payment/controller/PaymentController.java
@@ -23,30 +23,17 @@ public class PaymentController {
 
     private final PaymentService paymentService;
 
-    @PostMapping("/verify")
-    public BaseResponse<BaseResultDto> verifyPurchase(@RequestHeader(name = "Access-Member-Id") Long memberId,
-                                                      @RequestBody PurchaseRequest request) {
-
-        Long purchase = paymentService.purchase(memberId,
+    @PostMapping("/verify-and-consume")
+    public BaseResponse<BaseResultDto> verifyAndConsume(@RequestHeader(name = "Access-Member-Id") Long memberId,
+                                                   @RequestBody PurchaseRequest request) {
+        Long dataMemberId = paymentService.verifyProductPurchase(
+                memberId,
+                request.getType(),
                 request.getPackageName(),
                 request.getProductId(),
-                request.getPurchaseToken());
+                request.getPurchaseToken()
+        );
 
-        return BaseResponse.getSimpleRes(purchase);
-    }
-
-    @PostMapping("/verify-and-consume")
-    public ResponseEntity<String> verifyAndConsume(@RequestBody PurchaseRequest request) {
-        // 1. 구매 검증
-        ProductPurchase purchase = paymentService.verifyProductPurchase(request.getPackageName(), request.getProductId(), request.getPurchaseToken());
-
-        // 2. 구매 상태 확인
-        if (purchase.getPurchaseState() == 0 && purchase.getConsumptionState() == 0) {
-            // 3. 소비 처리
-            paymentService.consumeProduct(request.getPackageName(), request.getProductId(), request.getPurchaseToken());
-            return ResponseEntity.ok("Purchase verified and consumed successfully.");
-        }
-
-        return ResponseEntity.badRequest().body("Invalid or already consumed purchase.");
+        return BaseResponse.getSimpleRes(dataMemberId);
     }
 }
