+++ aiku/aiku-payment/src/main/java/payment/controller/dto/PurchaseRequest.java
@@ -1,5 +1,6 @@
 package payment.controller.dto;
 
+import common.domain.PaymentProductType;
 import jakarta.validation.constraints.NotNull;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -9,6 +10,8 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @NoArgsConstructor
 public class PurchaseRequest {
+    @NotNull
+    private PaymentProductType type;      // 앱의 패키지 이름
     @NotNull
     private String packageName;      // 앱의 패키지 이름
     @NotNull
