+++ aiku/aiku-payment/src/main/java/payment/service/PaymentService.java
@@ -1,10 +1,5 @@
 package payment.service;
 
-import com.google.api.client.googleapis.json.GoogleJsonResponseException;
-import com.google.api.client.http.HttpRequestInitializer;
-import com.google.api.client.http.javanet.NetHttpTransport;
-import com.google.api.client.json.gson.GsonFactory;
-import com.google.api.services.androidpublisher.AndroidPublisher;
 import com.google.api.services.androidpublisher.model.ProductPurchase;
 import common.domain.PaymentProduct;
 import common.domain.PaymentProductType;
@@ -16,85 +11,46 @@ import payment.exception.MemberNotFoundException;
 import payment.exception.PaymentException;
 import payment.repository.MemberRepository;
 import payment.repository.PaymentProductRepository;
-import payment.util.AccessTokenProvider;
-
-import java.io.IOException;
 
 @RequiredArgsConstructor
 @Transactional(readOnly = true)
 @Service
 public class PaymentService {
 
-    private final AccessTokenProvider accessTokenProvider;
     private final PaymentProductRepository paymentProductRepository;
     private final MemberRepository memberRepository;
-
-    private static final String APPLICATION_NAME = "com.example.myapp";
-
-    private AndroidPublisher initializePublisher() throws IOException {
-        String accessToken = accessTokenProvider.getAccessToken();
-
-        HttpRequestInitializer requestInitializer = request -> {
-            request.setConnectTimeout(60000);
-            request.setReadTimeout(60000);
-            request.getHeaders().setAuthorization("Bearer " + accessToken);
-        };
-
-        return new AndroidPublisher.Builder(new NetHttpTransport(), GsonFactory.getDefaultInstance(), requestInitializer)
-                .setApplicationName(APPLICATION_NAME)
-                .build();
-    }
+    private final GooglePaymentHelper googlePaymentHelper;
 
     @Transactional
-    public ProductPurchase verifyProductPurchase(Long memberId, PaymentProductType type, String packageName, String productId, String purchaseToken) {
-        try {
-            validateDB(purchaseToken);
-
-            PaymentProduct paymentProduct = getPaymentProduct(type);
-
-            makePayment(memberId, purchaseToken, paymentProduct);
-
-            AndroidPublisher publisher = initializePublisher();
-            return publisher.purchases().products()
-                    .get(packageName, productId, purchaseToken)
-                    .execute();
-        } catch (GoogleJsonResponseException e) {
-            System.err.println("GoogleJsonResponseException: " + e.getDetails());
-            throw new RuntimeException("Failed to verify purchase", e);
-        } catch (IOException e) {
-            System.err.println("IOException: " + e.getMessage());
-            throw new RuntimeException("Error initializing Google Play Publisher", e);
-        }
-    }
-
-    @Transactional
-    public void consumeProduct(PaymentProductType type, String packageName, String productId, String purchaseToken) {
-        try {
-            AndroidPublisher publisher = initializePublisher();
-            publisher.purchases().products()
-                    .consume(packageName, productId, purchaseToken)
-                    .execute();
-        } catch (GoogleJsonResponseException e) {
-            System.err.println("GoogleJsonResponseException: " + e.getDetails());
-            throw new RuntimeException("Failed to consume purchase", e);
-        } catch (IOException e) {
-            System.err.println("IOException: " + e.getMessage());
-            throw new RuntimeException("Error consuming purchase", e);
-        }
+    public Long verifyProductPurchase(Long memberId, PaymentProductType type, String packageName, String productId, String purchaseToken) {
+        validateDB(purchaseToken);
 
         PaymentProduct paymentProduct = getPaymentProduct(type);
+
+        makePayment(memberId, purchaseToken, paymentProduct);
+
+        ProductPurchase purchase = googlePaymentHelper.verifyProductPurchase(packageName, productId, purchaseToken);
+
+        if (purchase.getPurchaseState() == 0 && purchase.getConsumptionState() == 0) {
+            // 소비 처리
+            googlePaymentHelper.consumeProduct(packageName, productId, purchaseToken);
+            acceptPayment(paymentProduct, purchaseToken);
+        }
+
+        // TODO : 아쿠 증가 이벤트
+
+        return memberId;
+    }
+
+    private void acceptPayment(PaymentProduct paymentProduct, String purchaseToken) {
+        // DB Payment 상태 업데이트
         paymentProduct.acceptPayment(purchaseToken);
-
-        // 아쿠 증가 이벤트
-
-        // 아쿠 증가 실패 시 환불
-
-        // 충전 완료 알림
     }
 
     private PaymentProduct getPaymentProduct(PaymentProductType type) {
         PaymentProduct paymentProduct = paymentProductRepository.findByPaymentProductType(type)
                 .orElseThrow(PaymentException::new);
+
         return paymentProduct;
     }
 
@@ -108,10 +64,25 @@ public class PaymentService {
     }
 
     private void validateDB(String purchaseToken) {
-        // DB 내 purchaseToken 중복되는지 확인, true면 유일 보장
+        // DB 내 purchaseToken 중복되는지 확인, false -> 유일 보장
         if (paymentProductRepository.existsByPurchaseToken(purchaseToken)) {
             throw new PaymentException();
         }
+    }
+
+    // TODO : 보상 트랜잭션; 사용자 아쿠 증가 성공 시 알림
+    public void pointChargeSuccessAlarm() {
+
+    }
+
+    // TODO : 보상 트랜잭션; 사용자 아쿠 증가 실패 시 환불 처리
+    public void refund() {
+
+    }
+
+    // TODO : 보상 트랜잭션; 사용자 아쿠 증가 실패 시 실패 알림
+    public void pointChargeFailAlarm() {
+
     }
 
 }
